Make SetOverlap generate Long result. Fix JsonMigration version comparison.

+++ app/com/lynxanalytics/biggraph/graph_operations/SetOverlap.scala
@@ -19,7 +19,7 @@ object SetOverlap extends OpFromJson {
   class Output(implicit instance: MetaGraphOperationInstance, inputs: Input)
       extends MagicOutput(instance) {
     val overlaps = edgeBundle(inputs.segments.entity, inputs.segments.entity)
-    val overlapSize = edgeAttribute[Int](overlaps)
+    val overlapSize = edgeAttribute[Long](overlaps)
   }
   def fromJson(j: JsValue) = SetOverlap((j \ "minOverlap").as[Int])
 }
@@ -81,7 +81,7 @@ case class SetOverlap(minOverlap: Int) extends TypedMetaGraphOp[Input, Output] {
     val numberedEdgesWithOverlaps = edgesWithOverlaps.randomNumbered(rc.defaultPartitioner)
 
     output(o.overlaps, numberedEdgesWithOverlaps.mapValues(_._1))
-    output(o.overlapSize, numberedEdgesWithOverlaps.mapValues(_._2))
+    output(o.overlapSize, numberedEdgesWithOverlaps.mapValues(_._2.toLong))
   }
 
   // Generates the edges for a set of sets. This is O(n^2), but the set should
